requirements.txt
paho-mqtt
protobuf
pymongo